User and Group creation events made idempotent
We discovered that version 1.0.0 of the  user and group creation events were not-idempotent. This means that the system as a whole can not be restored from the event stream as each time the events are consumed the outcome would be different.

To be specific the events do not contain neither the user ids, nor the group ids. These ids are present in all the datastore events. In the event of a restore, users would lose access to all their files as the Gateway would submit group ids, from Heimdall, that are no longer recognised by the Datastore.

Fixing the situation for future groups and users is fairly easy. Instead of creating our ids when the events are consumed we add them to the event creation logic, thereby making them idempotent. The real problem lies in fixing the current groups and users.

For that issue, we propose a multi-stage solution:

1. Create new event formats, with new versions. This is done.
2. Submit events in the new format that would recreate the current state. They will be read and ignored by Heimdall.
3. Deploy the new version of Heimdall that only consumes the new
events.

This commit adds the new event versions and completely deprecates the
1.0.0 versions of the events. It also contains a 'patch' namespace
that will create the new 2.0.0 event versions, which on a restore will
return the system to it's current state. This will be manually
executed and one time only, and is included purely for informational purposes.