User data created on invite
When a user is invited the user data and self group data is
created. Users in this state are marked by the flag :pre-signup. These
users can be added to groups, kaylee namespace allows user and group
creation/joining, but they can not sign in. The signin protection is
based on the state flag and the nil password column.

State transistions are managed by events, but in keeping with Heimdall
current layout there are no Commands for these actions.

The invite failed event version has been incremented to improve the
format of that event.

The invite create event version has been incremented as it now
contains all the user information. Upon consumption the user and their
self group is created from this data.

Create user event remains the same, this should preserve replay as the
create event already had all the self group information.

The various persist functions have had conditions added to ensure
there isn't any overwriting of data.