Adding event constraint tests
There are two tests added.

The first, creates two users and adds them
to a shared group. The events for these actions tracked, and
resent. The test then checks that both the users appear in the groups
the correct number of times.

The second, creates the users and group again, then deletes all the
rows associated with them and resends the events. Checking that they
end up back where they were.

This gives us some confidents for both event replay and event idempotence.