Adding fdef and post assertion for signup-user
It didnt look worth introducing a spec check for signup-user fn,
because its very impure as used from the web handler.

Instead we introduced a small args test via fdef, and enabled
spec/instrument for the fn. We then added a post assertion for the
return type and enable asserts at test time in the cycle system fixture.