Improving add-member-to-group failure scenarios
Change introduces and or for the parition-key when sending failed to
add events, tries user, group ids, then gives up and creates a random
uid.

Change to kaylee name space to not attempt to add a member to group if
either group or user isn't found, improved error and success messages
to give more information.